feat: notifications for follow comments and likes

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -1,6 +1,8 @@
-import { validateRequest } from "@/auth";
-import prisma from "@/lib/prisma";
 import { LikeInfo } from "@/lib/types";
+
+import prisma from "@/lib/prisma";
+
+import { validateRequest } from "@/auth";
 
 export async function GET(
   req: Request,
@@ -40,6 +42,7 @@ export async function GET(
       likes: post._count.likes,
       isLikedByUser: !!post.likes.length,
     };
+
     return Response.json(data);
   } catch (error) {
     console.error(error);
@@ -55,33 +58,51 @@ export async function POST(
     const { user: loggedInUser } = await validateRequest();
 
     if (!loggedInUser) {
-      console.log("Unauthorized request: No user found.");
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    console.log("POST request to like post:");
-    console.log("Post ID:", postId);
-    console.log("Logged-in user ID:", loggedInUser.id);
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
+      },
+    });
 
-    await prisma.like.upsert({
-      where: {
-        userId_postId: {
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
+    await prisma.$transaction([
+      prisma.like.upsert({
+        where: {
+          userId_postId: {
+            userId: loggedInUser.id,
+            postId,
+          },
+        },
+        create: {
           userId: loggedInUser.id,
           postId,
         },
-      },
-      create: {
-        userId: loggedInUser.id,
-        postId,
-      },
-      update: {},
-    });
-
-    console.log("Like successfully added or already exists for post:", postId);
+        update: {},
+      }),
+      ...(loggedInUser.id !== post.userId
+        ? [
+            prisma.notification.create({
+              data: {
+                issuerId: loggedInUser.id,
+                recipientId: post.userId,
+                postId,
+                type: "LIKE",
+              },
+            }),
+          ]
+        : []),
+    ]);
 
     return new Response();
   } catch (error) {
-    console.error("Error in POST handler:", error);
+    console.error(error);
     return Response.json({ error: "Internal server error" }, { status: 500 });
   }
 }
@@ -97,12 +118,33 @@ export async function DELETE(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    await prisma.like.deleteMany({
-      where: {
-        userId: loggedInUser.id,
-        postId,
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
       },
     });
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
+    await prisma.$transaction([
+      prisma.like.deleteMany({
+        where: {
+          userId: loggedInUser.id,
+          postId,
+        },
+      }),
+      prisma.notification.deleteMany({
+        where: {
+          issuerId: loggedInUser.id,
+          recipientId: post.userId,
+          postId,
+          type: "LIKE",
+        },
+      }),
+    ]);
 
     return new Response();
   } catch (error) {
